producer_stg1/requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
cx-Oracle==7.0.0
idna==2.8
kafka-python==1.4.4
numpy==1.15.4
pandas==0.23.4
psycopg2==2.7.7
python-dateutil==2.7.5
pytz==2018.9
requests==2.21.0
six==1.12.0
soupsieve==1.6.2
SQLAlchemy==1.2.15
urllib3==1.24.1
lxml==4.2.5
cx-Oracle==7.0.0
avro-python3==1.8.2
